+++ src/format/pg/parser.js
@@ -287,29 +287,37 @@ function peg$parse(input, options) {
     edges.push(e)
  };
   var peg$f3 = function(id) { return id };
-  var peg$f4 = function(name, value) {
+  var peg$f4 = function(id) { 
+      if (id === "") { error("Identifiers cannot be empty") }
+      return id
+    };
+  var peg$f5 = function(name, value) {
       return [ name, value ]
     };
-  var peg$f5 = function(name) {
+  var peg$f6 = function(key) {
+        if (key === "") { error("Property keys cannot be empty") }
+        return key
+      };
+  var peg$f7 = function(name) {
       return name.join("").slice(0,-1)
     };
-  var peg$f6 = function(first, rest) {
+  var peg$f8 = function(first, rest) {
       return [first, ...rest]
     };
-  var peg$f7 = function() { return true };
-  var peg$f8 = function() { return false };
-  var peg$f9 = function() { return null };
-  var peg$f10 = function(chars) { return chars.join("") };
-  var peg$f11 = function() { return "\b" };
-  var peg$f12 = function() { return "\f" };
-  var peg$f13 = function() { return "\n" };
-  var peg$f14 = function() { return "\r" };
-  var peg$f15 = function() { return "\t" };
-  var peg$f16 = function(sequence) { return sequence };
-  var peg$f17 = function(digits) {
+  var peg$f9 = function() { return true };
+  var peg$f10 = function() { return false };
+  var peg$f11 = function() { return null };
+  var peg$f12 = function(chars) { return chars.join("") };
+  var peg$f13 = function() { return "\b" };
+  var peg$f14 = function() { return "\f" };
+  var peg$f15 = function() { return "\n" };
+  var peg$f16 = function() { return "\r" };
+  var peg$f17 = function() { return "\t" };
+  var peg$f18 = function(sequence) { return sequence };
+  var peg$f19 = function(digits) {
       return String.fromCharCode(parseInt(digits, 16))
     };
-  var peg$f18 = function() { return parseFloat(text()) };
+  var peg$f20 = function() { return parseFloat(text()) };
   var peg$currPos = options.peg$currPos | 0;
   var peg$savedPos = peg$currPos;
   var peg$posDetailsCache = [{ line: 1, column: 1 }];
@@ -942,9 +950,15 @@ function peg$parse(input, options) {
   }
 
   function peg$parseIdentifier() {
-    var s0;
+    var s0, s1;
 
-    s0 = peg$parseQuotedString();
+    s0 = peg$currPos;
+    s1 = peg$parseQuotedString();
+    if (s1 !== peg$FAILED) {
+      peg$savedPos = s0;
+      s1 = peg$f4(s1);
+    }
+    s0 = s1;
     if (s0 === peg$FAILED) {
       s0 = peg$parseUnquotedIdentifier();
     }
@@ -1043,7 +1057,7 @@ function peg$parse(input, options) {
         s3 = peg$parseValueList();
         if (s3 !== peg$FAILED) {
           peg$savedPos = s0;
-          s0 = peg$f4(s2, s3);
+          s0 = peg$f5(s2, s3);
         } else {
           peg$currPos = s0;
           s0 = peg$FAILED;
@@ -1069,25 +1083,33 @@ function peg$parse(input, options) {
     var s0, s1, s2, s3, s4, s5, s6, s7, s8, s9;
 
     s0 = peg$currPos;
-    s1 = peg$parseQuotedString();
-    if (s1 !== peg$FAILED) {
-      s2 = peg$parseSpace();
-      if (s2 === peg$FAILED) {
-        s2 = null;
+    s1 = peg$currPos;
+    s2 = peg$parseQuotedString();
+    if (s2 !== peg$FAILED) {
+      s3 = peg$parseSpace();
+      if (s3 === peg$FAILED) {
+        s3 = null;
       }
       if (input.charCodeAt(peg$currPos) === 58) {
-        s3 = peg$c5;
+        s4 = peg$c5;
         peg$currPos++;
       } else {
-        s3 = peg$FAILED;
+        s4 = peg$FAILED;
         if (peg$silentFails === 0) { peg$fail(peg$e13); }
       }
-      if (s3 !== peg$FAILED) {
-        s0 = s1;
+      if (s4 !== peg$FAILED) {
+        peg$savedPos = s1;
+        s1 = peg$f6(s2);
       } else {
-        peg$currPos = s0;
-        s0 = peg$FAILED;
+        peg$currPos = s1;
+        s1 = peg$FAILED;
       }
+    } else {
+      peg$currPos = s1;
+      s1 = peg$FAILED;
+    }
+    if (s1 !== peg$FAILED) {
+      s0 = s1;
     } else {
       peg$currPos = s0;
       s0 = peg$FAILED;
@@ -1315,7 +1337,7 @@ function peg$parse(input, options) {
         }
         if (s1 !== peg$FAILED) {
           peg$savedPos = s0;
-          s1 = peg$f5(s1);
+          s1 = peg$f7(s1);
         }
         s0 = s1;
       }
@@ -1402,7 +1424,7 @@ function peg$parse(input, options) {
         }
       }
       peg$savedPos = s0;
-      s0 = peg$f6(s1, s2);
+      s0 = peg$f8(s1, s2);
     } else {
       peg$currPos = s0;
       s0 = peg$FAILED;
@@ -1531,7 +1553,7 @@ function peg$parse(input, options) {
         }
         if (s1 !== peg$FAILED) {
           peg$savedPos = s0;
-          s1 = peg$f7();
+          s1 = peg$f9();
         }
         s0 = s1;
         if (s0 === peg$FAILED) {
@@ -1545,7 +1567,7 @@ function peg$parse(input, options) {
           }
           if (s1 !== peg$FAILED) {
             peg$savedPos = s0;
-            s1 = peg$f8();
+            s1 = peg$f10();
           }
           s0 = s1;
           if (s0 === peg$FAILED) {
@@ -1559,7 +1581,7 @@ function peg$parse(input, options) {
             }
             if (s1 !== peg$FAILED) {
               peg$savedPos = s0;
-              s1 = peg$f9();
+              s1 = peg$f11();
             }
             s0 = s1;
           }
@@ -1597,7 +1619,7 @@ function peg$parse(input, options) {
       }
       if (s3 !== peg$FAILED) {
         peg$savedPos = s0;
-        s0 = peg$f10(s2);
+        s0 = peg$f12(s2);
       } else {
         peg$currPos = s0;
         s0 = peg$FAILED;
@@ -1651,7 +1673,7 @@ function peg$parse(input, options) {
         }
         if (s3 !== peg$FAILED) {
           peg$savedPos = s2;
-          s3 = peg$f11();
+          s3 = peg$f13();
         }
         s2 = s3;
         if (s2 === peg$FAILED) {
@@ -1665,7 +1687,7 @@ function peg$parse(input, options) {
           }
           if (s3 !== peg$FAILED) {
             peg$savedPos = s2;
-            s3 = peg$f12();
+            s3 = peg$f14();
           }
           s2 = s3;
           if (s2 === peg$FAILED) {
@@ -1679,7 +1701,7 @@ function peg$parse(input, options) {
             }
             if (s3 !== peg$FAILED) {
               peg$savedPos = s2;
-              s3 = peg$f13();
+              s3 = peg$f15();
             }
             s2 = s3;
             if (s2 === peg$FAILED) {
@@ -1693,7 +1715,7 @@ function peg$parse(input, options) {
               }
               if (s3 !== peg$FAILED) {
                 peg$savedPos = s2;
-                s3 = peg$f14();
+                s3 = peg$f16();
               }
               s2 = s3;
               if (s2 === peg$FAILED) {
@@ -1707,7 +1729,7 @@ function peg$parse(input, options) {
                 }
                 if (s3 !== peg$FAILED) {
                   peg$savedPos = s2;
-                  s3 = peg$f15();
+                  s3 = peg$f17();
                 }
                 s2 = s3;
                 if (s2 === peg$FAILED) {
@@ -1739,7 +1761,7 @@ function peg$parse(input, options) {
       }
       if (s2 !== peg$FAILED) {
         peg$savedPos = s0;
-        s0 = peg$f16(s2);
+        s0 = peg$f18(s2);
       } else {
         peg$currPos = s0;
         s0 = peg$FAILED;
@@ -1781,7 +1803,7 @@ function peg$parse(input, options) {
     }
     if (s1 !== peg$FAILED) {
       peg$savedPos = s0;
-      s1 = peg$f17(s1);
+      s1 = peg$f19(s1);
     }
     s0 = s1;
 
@@ -1827,7 +1849,7 @@ function peg$parse(input, options) {
         s4 = null;
       }
       peg$savedPos = s0;
-      s0 = peg$f18();
+      s0 = peg$f20();
     } else {
       peg$currPos = s0;
       s0 = peg$FAILED;